Reproducible secure seeds and salts for hashing
- Created a ReproducibleSeedGenerator that produces a secure seed value
for a piece of associated data, seeded with a common secret seed value.
The generated seed is unpredictable, yet can be reproduced given these
two pieces of information.

- Created a ReproducibleSaltGenerator that produces a sequence of
unpredictable salt values for use in salted hashing, which are generated
based on a given piece of associated data and a common secret seed
value.

- The ReproducibleSeedGenerator is used to produce a secret seed value
associated with this piece of data, which is then used to seed a
Deterministic Random Bit Generator (DRBG) that is used to generate the
sequence of salt values.
  By seeding this DRBG in a deterministic way, based on the associated
data and common secret seed, this process is reproducible making it
possible to regenerate the same sequence of salts with possession of
only these two pieces of information.

- Created a Maven pom.xml for this new project. Added dependencies for
Apache Commons Codec & Crypto.
  Configured the maven-compiler-plugin for Java SE 11. Configured the
maven-surefire-plugin with support for the needed reflection access into
Java 11 modules from test frameworks.