Salted Hash Generator & MemoryUtils
- Created a salted hash generator that uses a secret seed, together with
ReproducibleSaltGenerator, to provide a Iterator for iterating over the
salted hashes for a given cleartext value.
  For a given secret seed, the Iterator produced for a given cleartext
value will always return the same sequence of salted hashes.

- Created a MemoryUtils utility class, providing functions for
concatenating an array of byte-array buffers and another for clearing a
sensitive byte-array buffer (with zeros).

- Updated HashingUtils to use the buffer concatenating feature of
MemoryUtils, instead of manually concatenating the cleartext and salt
byte-arrays when preparing to create the (SHA-256) hash, using
System.arraycopy(...).

- Created an application.yml configuration file, adding a
'securerandom.drbg.config' property configuring the DRBG associated with
a SecureRandom created with SecureRandom algorithm "DRBG".
  Configured the DRBG to use Hash_DRBG with SHA-256 and a security
strength of 128 bits, sufficient for production of 16 B salt values.